@@ -3,17 +3,45 @@
   "version": "1.2.0",
   "private": false,
   "description": "A Vue.js component to create beautiful animated circular progress bars",
-  "main": "./dist/vue-ellipse-progress.umd.min.js",
+  "author": {
+    "name": "Sergej Atamantschuk",
+    "email": "sergejatam@gmail.com"
+  },
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build --target lib --formats umd-min --name vue-ellipse-progress ./src/plugin.js",
     "lint": "vue-cli-service lint --fix",
-    "test": "vue-cli-service test:unit --colors",
-    "publish-beta": "npm run lint && npm run build && npm publish --tag beta"
+    "publish-beta": "npm run lint && npm run build && npm publish --tag beta",
+    "test": "vue-cli-service test:unit --colors"
   },
-  "author": {
-    "name": "Sergej Atamantschuk",
-    "email": "sergejatam@gmail.com"
+  "main": "./dist/vue-ellipse-progress.umd.min.js",
+  "files": [
+    "dist",
+    "README"
+  ],
+  "devDependencies": {
+    "@vue/cli-plugin-babel": "^4.5.4",
+    "@vue/cli-plugin-eslint": "^4.5.4",
+    "@vue/cli-plugin-unit-mocha": "^4.5.4",
+    "@vue/cli-service": "^4.5.4",
+    "@vue/compiler-sfc": "^3.0.0-beta.1",
+    "@vue/eslint-config-airbnb": "^5.1.0",
+    "@vue/test-utils": "^2.0.0-alpha.1",
+    "babel-eslint": "^10.1.0",
+    "chai": "^4.2.0",
+    "eslint": "^7.7.0",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-prettier": "^3.1.4",
+    "eslint-plugin-vue": "^7.0.0-alpha.0",
+    "node-sass": "^4.14.1",
+    "prettier": "^2.0.5",
+    "sass-loader": "^9.0.3",
+    "@vue/cli": "^4.5.4",
+    "vue": "^3.0.0-beta.1",
+    "vue-cli-plugin-vue-next": "~0.1.3"
+  },
+  "peerDependencies": {
+    "vue": "^3"
   },
   "bugs": "https://github.com/setaman/vue-ellipse-progress/issues",
   "homepage": "https://vue-ellipse-progress-demo.netlify.app/#/",
@@ -29,31 +57,5 @@
     "progress",
     "bar"
   ],
-  "repository": "github:setaman/vue-ellipse-progress",
-  "peerDependencies": {
-    "vue": "^2"
-  },
-  "devDependencies": {
-    "@vue/cli-plugin-babel": "^4.5.4",
-    "@vue/cli-plugin-eslint": "^4.5.4",
-    "@vue/cli-plugin-unit-mocha": "^4.5.4",
-    "@vue/cli-service": "^4.5.4",
-    "@vue/eslint-config-airbnb": "^5.1.0",
-    "@vue/test-utils": "1.0.4",
-    "babel-eslint": "^10.1.0",
-    "chai": "^4.2.0",
-    "eslint": "^7.7.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-vue": "^6.2.2",
-    "node-sass": "^4.14.1",
-    "prettier": "^2.0.5",
-    "sass-loader": "^9.0.3",
-    "vue": "^2.6.12",
-    "vue-template-compiler": "^2.6.12"
-  },
-  "files": [
-    "dist",
-    "README"
-  ]
+  "repository": "github:setaman/vue-ellipse-progress"
 }
